removed and moved more files
