+++ src/popup/index.html
@@ -8,8 +8,8 @@
         height: 400px;
       }
     </style>
-    <link rel="stylesheet" href="../css/popup.css" />
-    <link rel="stylesheet" href="../css/ErrorPopupStyle.css"/>
+    <link rel="stylesheet" href="./popup.css" />
+    <link rel="stylesheet" href="../components/ErrorPopup/ErrorPopupStyle.css"/>
   </head>
 
   <body>
